requirements.txt
matplotlib
numpy
pyserial
scipy
